+++ lib/Question.dart
@@ -1,0 +1,11 @@
+class Question {
+  final String questionText;
+  final List<String> options;
+  final String correctOption;
+
+  Question({
+    required this.questionText,
+    required this.options,
+    required this.correctOption,
+  });
+}
